requirements.txt
streamlit
easyocr
google-genai
pillow